Work-in-progress change to allow profiling of node.js 0.12.
The V8 profiler API changed greatly between V8 3.14 (in node 0.10)
and 3.28 (in node 0.12.0):
- a few functions got renamed
- quite a few functions were removed entirely
- the remaining CpuProfile and HeapProfile methods were mostly
  static and now require an object instance, to deal with the
  V8 Isolate (which is also where you get the object instance)

I just #ifdef'd out references to the functions that don't exist
any more, in the hopes that the frontend wouldn't call them, and
I just use the current Isolate anywhere an instance is needed.
This seems to mostly work.

Current status:
- node 0.10: this compiles, seems to work
- node 0.12: this compiles, seems to mostly work.
  - heap snapshot seems to work
  - missing bottom-up cpu profile is annoying because UI starts
    with that one by default
  - even in top-down cpu profile all the samples are 0