A few more fixes for Node 0.12's newer V8.
Invoke StartProfiling with record_samples=true.

Expose the new CpuProfileNode::GetHitCount accessor, even though I'm
not sure whether we can get this exposed in the UI.

Add a few more comments on the differences in properties/accessors
exposed.

The V26 frontend still doesn't really know what to do with the CPU
profiles we feed it, and show a call graph with 0 time spent.